+++ src/vi/k2600.ttm/legacy/Legacy/IColdResistanceConfig.cs
@@ -12,26 +12,26 @@ public interface IColdResistanceConfig : ICloneable
     /// <summary> Gets or sets the integration period in number of power line cycles for measuring the
     /// cold resistance. </summary>
     /// <value> The aperture. </value>
-    float Aperture { get; set; }
+    public float Aperture { get; set; }
 
     /// <summary> Gets or sets the current level to apply to the device for measuring the cold
     /// resistance. </summary>
     /// <value> The current level. </value>
-    float CurrentLevel { get; set; }
+    public float CurrentLevel { get; set; }
 
     /// <summary> Gets or sets the high limit for determining measurement pass fail condition. </summary>
     /// <value> The high limit. </value>
-    float HighLimit { get; set; }
+    public float HighLimit { get; set; }
 
     /// <summary> Gets or sets the low limit for determining measurement pass fail condition. </summary>
     /// <value> The low limit. </value>
-    float LowLimit { get; set; }
+    public float LowLimit { get; set; }
 
     /// <summary> Restores known state. </summary>
-    void ResetKnownState();
+    public void ResetKnownState();
 
     /// <summary> Gets or sets the voltage limit. </summary>
     /// <value> The voltage limit. </value>
-    float VoltageLimit { get; set; }
+    public float VoltageLimit { get; set; }
 
 }